(PDB-4832) Add test for migration 76 in isolation
This commit tests the case where the old version of migration 74 has
already been applied and the idx_report_id index wasn't added to the
partitions. In this case migration 76 will add the index to any existing
partitions. Changing the behavior of migration 74 after the fact should
be safe because the index creation is guarded by 'if not exists' in both
migration 74 and 76. If we've already migrated past 74 before the change
then the index will get added in 76. If someone applies both 74 and 76
at the same time 76 will then become a no-op.